Reintroduce `vnet/install_service_windows.go`
